Mudanças em algumas rotas e padronização parcial do código

+++ src/routes/diretores.js
@@ -1,7 +1,5 @@
 const express = require("express");
-
 const router = express.Router();
-
 const db = require("../db");
 
 function verificarId(id) {
@@ -14,34 +12,24 @@ function verificarId(id) {
     return true;
 }
 
-
-// GET - todos os diretores
+//Get - todos os diretores
 router.get("/", async (req, res) => {
     try {
-        const resultado = await db.query(
-            "SELECT * FROM diretores"
-        );
+        const resultado = await db.query("SELECT * FROM diretores");
+        if (resultado.rowCount == 0){
+            return res.status(404).json({msg: "Não há nenhum diretor no sistema!"})
+        }
 
         res.status(200).json({diretores: resultado.rows});
-
     } catch (erro) {
-        res.status(500).json({
-            erro: erro.message
-        });
+        res.status(500).json({msg: erro});
     }
 });
 
-
-
-//-----------------------
-//Diretores/dirige
-//-------------------------
-
-
-
+//Get - todos os filmes que cada diretor dirigiu
 router.get("/dirige", async (req, res)=>{
     try{
-        const r = await db.query(`SELECT * FROM diretores JOIN filmes ON diretores.id = filmes.diretor`);
+        const r = await db.query("SELECT * FROM diretores JOIN filmes ON diretores.id = filmes.diretor");
         if (r.rowCount==0){
             return res.status(404).json("Bixou, nada encontrado")
         }
@@ -71,22 +59,20 @@ router.get("/dirige", async (req, res)=>{
 
         res.json(l2)
     }catch(erro){
-        res.status(400).json("Bixou, "+erro)
+        res.status(400).json({msg: "Bixou, "+erro})
     }
 })
 
-router.get("/dirige/:id", async (req, res) => {
+//Get - todos os filmes de um determinado diretor
+router.get("/:id/dirige/", async (req, res) => {
     try {
         if (!verificarId(req.params.id)) {
-            return res.status(400).json({
-                erro: "O id deve ser um número válido"
-            });
+            return res.status(400).json({msg: "O id deve ser um número válido"});
         }
 
         const r = await db.query(`SELECT * FROM filmes JOIN diretores ON filmes.diretor = diretores.id WHERE diretores.id = $1`, [req.params.id]);
-
         if (r.rowCount === 0) {
-            return res.status(404).json("Bixou, nenhum filme encontrado para esse diretor ou o diretor não existe");
+            return res.status(404).json({msg: "Bixou, nenhum filme encontrado para esse diretor ou o diretor não existe"});
         }
 
         let dic={[r.rows[0].nome]:[]}
@@ -96,10 +82,11 @@ router.get("/dirige/:id", async (req, res) => {
         }
         res.json(dic);
     } catch (erro) {
-        res.status(400).json("Bixou, " + erro);
+        res.status(400).json({msg: "Bixou, " + erro});
     }
 });
 
+//Get - os 3 diretores mais premiados
 router.get("/premiacoes", async (req, res)=>{
     try{
         const r=await db.query("SELECT * FROM diretores");
@@ -128,65 +115,43 @@ router.get("/premiacoes", async (req, res)=>{
     }
 });
 
-
-// GET - diretor por ID
+//Get - diretor específico
 router.get("/:id", async (req, res) => {
     try {
         const { id } = req.params;
         if (!verificarId(id)) {
-            return res.status(400).json({
-                erro: "O id deve ser um número válido"
-            });
+            return res.status(400).json({msg: "O id deve ser um número válido"});
         }
 
-        const resultado = await db.query(
-            "SELECT * FROM diretores WHERE id = $1",
-            [id]
-        );
-
+        const resultado = await db.query("SELECT * FROM diretores WHERE id = $1", [id]);
         if (resultado.rows.length === 0) {
-            return res.status(404).json({
-                msg: "Diretor não encontrado"
-            });
+            return res.status(404).json({msg: "Diretor não encontrado"});
         }
 
         res.status(200).json({diretor: resultado.rows[0]});
-
     } catch (erro) {
-        res.status(500).json({
-            erro: erro.message
-        });
+        res.status(500).json({msg:"Bixou"+erro});
     }
 });
 
+//Post - criar um novo diretor
 router.post("/", async (req, res) => {
     try {
         const { nome, nascimento, descricao, qtde_premios } = req.body;
 
-        const resultado = await db.query(
-            `INSERT INTO diretores 
-            (nome, nascimento, descricao, qtde_premios)
-            VALUES ($1, $2, $3, $4)
-            RETURNING *`,
-            [nome, nascimento, descricao, qtde_premios]
-        );
-
-        res.status(201).json({msg: "Diretor adicionado com sucesso!", diretor: resultado.rows[0]});
+        const resultado = await db.query("INSERT INTO diretores (nome, nascimento, descricao, qtde_premios) VALUES ($1, $2, $3, $4) RETURNING *", [nome, nascimento, descricao, qtde_premios]);
+        res.status(200).json({msg: "Diretor adicionado com sucesso!", diretor: resultado.rows[0]});
 
     } catch (erro) {
-        res.status(500).json({
-            erro: erro.message
-        });
+        res.status(500).json({msg: erro});
     }
 });
 
+//Delete - deletar um diretor se ele não tiver nenhum filme associado
 router.delete("/:id", async (req, res) => {
     try {
-        const { id } = req.params;
-        if (!verificarId(id)) {
-            return res.status(400).json({
-                erro: "O id deve ser um número válido"
-            });
+        if (!verificarId(req.params.id)) {
+            return res.status(400).json({msg: "O id deve ser um número válido"});
         }
 
         const { idUsuario } = req.body || {}
@@ -195,9 +160,13 @@ router.delete("/:id", async (req, res) => {
             return res.status(400).json({msg: "Usuários não existente ou não é administrador!"})
         }
 
-        const resultado = await db.query("DELETE FROM diretores WHERE id = $1 RETURNING *", [id]);
-
-        if (resultado.rows.length === 0) {
+        const r2 = await db.query("SELECT diretores.id diretor, filmes.id filme FROM diretores JOIN filmes ON diretores.id = filmes.diretor WHERE diretores.id = $1", [req.params.id])
+        if (r2.rowCount == 0){
+            const resultado = await db.query("DELETE FROM diretores WHERE id = $1 RETURNING *", [id]);
+        }else{
+            return res.status(400).json({msg: "Este diretor tem filme(s) associados, não é possível deletá-lo!"})
+        }
+        /*if (resultado.rows.length === 0) {
             return res.status(404).json({
                 msg: "Diretor não encontrado"
             });
@@ -206,12 +175,11 @@ router.delete("/:id", async (req, res) => {
         res.status(200).json({
             msg: "Diretor deletado com sucesso",
             diretor: resultado.rows[0]
-        });
+        });*/
+        res.json({msg: "Diretor deletado com sucesso!"})
 
     } catch (erro) {
-        res.status(500).json({
-            erro: erro.message
-        });
+        res.status(500).json({msg: erro});
     }
 });
 
@@ -220,43 +188,39 @@ router.put("/:id", async (req, res) => {
         const { id } = req.params;
         const { idUsuario, nome, nascimento, descricao, qtde_premios } = req.body;
 
+        //Verificações sobre a entrada:
+            if (nome.length > 85){
+                return res.status(400).json({msg: "Nome ultrapassou os limites de caracteres!"})
+            }
+            if (!verificarId(qtde_premios)) {
+                return res.status(400).json({msg: "A quantidade de prêmios deve ser um número válido"});
+            }
+
         if (!verificarId(id)) {
-            return res.status(400).json({
-                erro: "O id deve ser um número válido"
-            });
+            return res.status(400).json({msg: "O id deve ser um número válido"});
         }
         const r1 = await db.query("SELECT id FROM usuarios WHERE id = $1 AND administrador = TRUE", [idUsuario])
         if (r1.rows.length === 0){
-            return res.status(400).json({msg: "Usuários não existente ou não é administrador!"})
+            return res.status(400).json({msg: "Usuário não existente ou não é administrador!"})
         }
 
+        const anterior = await db.query("SELECT * FROM diretores WHERE id=$1", [id])
+        if (anterior.rowCount == 0){
+            return res.status(400).json({msg: "Não existe diretor com esse id!"})
+        }
+        const novoNome = nome ?? diretorAtual.nome;
+        const novoNascimento = nascimento ?? diretorAtual.nascimento;
+        const novaDescricao = descricao ?? diretorAtual.descricao;
+        const novaQtdePremios = qtde_premios ?? diretorAtual.qtde_premios;
 
-        const resultado = await db.query(
-            `UPDATE diretores
-             SET nome = $1,
-                 nascimento = $2,
-                 descricao = $3,
-                 qtde_premios = $4
-             WHERE id = $5
-             RETURNING *`,
-            [nome, nascimento, descricao, qtde_premios, id]
-        );
-
+        const resultado = await db.query("UPDATE diretores SET nome = $1, nascimento = $2, descricao = $3, qtde_premios = $4 WHERE id = $5 RETURNING *", [novoNome, novoNascimento, novaDescricao, novaQtdePremios, id]);
         if (resultado.rows.length === 0) {
-            return res.status(404).json({
-                msg: "Diretor não encontrado"
-            });
+            return res.status(404).json({msg: "Erro ao atualizar os dados!"});
         }
 
-        res.status(200).json({
-            msg: "Diretor atualizado com sucesso",
-            diretor: resultado.rows[0]
-        });
-
+        res.status(200).json({msg: "Diretor atualizado com sucesso", diretor: resultado.rows[0] });
     } catch (erro) {
-        res.status(500).json({
-            erro: erro.message
-        });
+        res.status(500).json({msg: erro});
     }
 });
